+++ src/main/java/com/crystalpizaa/api/service/PizzaServiceImpl.java
@@ -1,13 +1,17 @@
 package com.crystalpizaa.api.service;
 
+import com.crystalpizaa.api.AppLogger;
 import com.crystalpizaa.api.dao.interfaces.PizzaRepository;
 import com.crystalpizaa.api.service.models.core.Pizza;
+import com.crystalpizaa.api.service.models.validation.DataNotFoundException;
 import com.crystalpizaa.api.service.models.validation.ErrorInfo;
 import com.crystalpizaa.api.service.models.validation.ValidationException;
 import com.crystalpizaa.api.service.models.validation.ValidationInfo;
 import com.crystalpizaa.api.service.translators.PizzaTranslator;
 import java.util.ArrayList;
 import java.util.List;
+import javax.persistence.EntityNotFoundException;
+import org.slf4j.event.Level;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,14 +22,23 @@ public class PizzaServiceImpl implements PizzaService {
   private PizzaRepository pizzaRepository;
 
   @Override
-  public Pizza save(Pizza pizza)  {
+  public Pizza save(Pizza pizza) {
 
     if (EnsureValid(pizza)) {
-      com.crystalpizaa.api.dao.entities.Pizza pizzaObj = PizzaTranslator.ToDaoModel(pizza);
+      try {
+        com.crystalpizaa.api.dao.entities.Pizza pizzaObj = PizzaTranslator.ToDaoModel(pizza);
 
-      com.crystalpizaa.api.dao.entities.Pizza result = this.pizzaRepository.save(pizzaObj);
+        com.crystalpizaa.api.dao.entities.Pizza result = this.pizzaRepository.save(pizzaObj);
 
-      return PizzaTranslator.ToServiceModel(result);
+        return PizzaTranslator.ToServiceModel(result);
+      } catch (Exception ex) {
+        if (ex instanceof ValidationException) {
+          throw ex;
+        }
+        AppLogger
+            .log(this.getClass().getName(), "Exception occurred:" + ex.getMessage(), Level.ERROR);
+      }
+
     }
     return null;
   }
@@ -33,24 +46,80 @@ public class PizzaServiceImpl implements PizzaService {
   @Override
   public List<Pizza> getAll() {
 
-    List<com.crystalpizaa.api.dao.entities.Pizza> pizzaObjs = this.pizzaRepository.findAll();
-
-    return PizzaTranslator.ToServiceModel(pizzaObjs);
+    try {
+      List<com.crystalpizaa.api.dao.entities.Pizza> pizzaObjs = this.pizzaRepository.findAll();
+      if (pizzaObjs == null || pizzaObjs.size() == 0) {
+        ErrorInfo error = new ErrorInfo();
+        error.setTitle("Data not found");
+        error.setDetails("Pizzas not found");
+        AppLogger.log(this.getClass().getName(), "Pizzas not found", Level.INFO);
+        throw new DataNotFoundException(error);
+      }
+      return PizzaTranslator.ToServiceModel(pizzaObjs);
+    } catch (Exception ex) {
+      if (ex instanceof DataNotFoundException) {
+        throw ex;
+      }
+      AppLogger
+          .log(this.getClass().getName(), "Exception occurred:" + ex.getMessage(), Level.ERROR);
+    }
+    return null;
   }
 
   @Override
   public Pizza get(int id) {
 
-    com.crystalpizaa.api.dao.entities.Pizza pizzaObj = this.pizzaRepository.getOne(id);
+    try {
 
-    return PizzaTranslator.ToServiceModel(pizzaObj);
+      com.crystalpizaa.api.dao.entities.Pizza pizzaObj = this.pizzaRepository.findById(id)
+          .orElse(null);
+      if (pizzaObj == null) {
+        ErrorInfo error = new ErrorInfo();
+        error.setTitle("Data not found");
+        error.setDetails(String.format("Requested Pizza of id %d not found", id));
+        AppLogger
+            .log(this.getClass().getName(), String.format("Requested Pizza of id %d not found", id),
+                Level.INFO);
+        throw new DataNotFoundException(error);
+      }
+
+      return PizzaTranslator.ToServiceModel(pizzaObj);
+    } catch (Exception ex) {
+      if (ex instanceof DataNotFoundException) {
+        throw ex;
+      }
+      AppLogger
+          .log(this.getClass().getName(), "Exception occurred:" + ex.getMessage(), Level.ERROR);
+    }
+
+    return null;
   }
 
   @Override
   public boolean remove(int id) {
-    this.pizzaRepository.deleteById(id);
+    try {
+      com.crystalpizaa.api.dao.entities.Pizza pizzaObj = this.pizzaRepository.findById(id)
+          .orElse(null);
+      if (pizzaObj == null) {
+        ErrorInfo error = new ErrorInfo();
+        error.setTitle("Data not found");
+        error.setDetails(String.format("Requested pizza of id %d not found", id));
+        AppLogger
+            .log(this.getClass().getName(), String.format("Requested pizza of id %d not found", id),
+                Level.INFO);
+        throw new DataNotFoundException(error);
+      }
+      this.pizzaRepository.deleteById(id);
+      return true;
+    } catch (Exception ex) {
+      if (ex instanceof DataNotFoundException) {
+        throw ex;
+      }
+      AppLogger
+          .log(this.getClass().getName(), "Exception occurred:" + ex.getMessage(), Level.ERROR);
+    }
 
-    return true;
+    return false;
   }
 
   private boolean EnsureValid(Pizza pizza) {
